libglusterfsclient: Perform explicit create
We must create the file based on the O_CREATE in the flags
instead of depending on the return status of the lookup.
This works because we've already handled O_EXCL earlier and
also handled a missing file(ENOENT)
in case this was actually an open and not create.
Adapted from Raghu's original patch.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>