libglusterfsclient: Add virtual mount point support
One part of the glusterfsclient library requires users
to use a handle to address a particular glusterfs context.
This can become cumbersome for applications that need to use multiple
such context. For such applications, we make it easy by letting them
define virtual mount points(VMPs), which are just paths that identify
a particular glusterfs context. This is done through a new
glusterfs_mount(..) interface.

Subsequenly, any path based operations, that pre-pend a VMP before
the files and directories inside the glusterfsd export, get internally
routed to the correct glusterfs context. This helps users do away
with the need to maintain their own path-to-glusterfs_handle_t
mapping.

The old handle-based interface still exists for whoever wishes to use
it.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>