io-threads: Clarify thread count range
This patch cleans up io-threads behaviour regarding the
range values that can be specified for min-threads
and max-threads. THe major change is that the min threads
have been reduced to 2 to signify that io-threads needs minimum
two threads for its operation, while keeping the default number of
threads at 16. The idea is to decouple the default thread count
from the minimum thread count.

Note to Avati:
This applies over Raghu's indentation and logging take-3 patch.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>