libglusterfsclient: Add comments to API
Some of the functions have also been moved within
the header to reflect the order in which users will
read through and try to understand how to use the lib.

For eg. the init, mount and umount functions have been moved to
the beginning to clarify how to start by init and de-initing.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>