libglusterfsclient: Fix realpath operation
This commit does two things:
1. Ensures we go on to traverse nested symlinks by calling realpath
recursively on the symlink.

2. Makes glusterfs_realpath, a VMP-based operation, return a resolved
/real path that actually has the VMP pre-fixed to the resolved path.
If this is not done, we'll be returning a path that is valid only
within the glusterfs context, i.e. the path will point to a real
file/dir only if a glusterfs handle identifies which glusterfs context
the file is in.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>