booster: Add VMP-based open support
In brief, when an open is called the VMP-based open is called first.
If that fails, the fallback is the Mount-Point Bypass(MBP) technique
which uses entended attributes to determine if the given
path is a gluster mount. If it is, it goes ahead to create a
glusterfsclient handle. All subsequent calls to this real mount point
get redirected to the glusterfsclient handle. The other fallback is
the actual open syscall on the local file system itself. This happens
when the absence of an extended attribute tells us that the path is
not even a glusterfs mount point.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>