mem-pool: Restructure mem-pool behaviour
This commit changes mem-pool behaviour to return a directly usable
address by performing the required adjustment on the address
being returned.

This is different from the previous behaviour where we're trying to fit
into the requested size, the list_head*2 also. This is not efficient
enough in terms of space but hopefully works better than not having any
mem-pool at all. Besides, I am not comfortable with mem-pool meta-data
and caller-useable memory area being the same because of the potential for
mem-pool's data structure corruption.

PS:
Please do read the comments in the code for more info during review.

Signed-off-by: Anand V. Avati <avati@amp.gluster.com>