booster: Eliminate gluster context creation race
When multiple threads try to create a glusterfs context using the
glusterfs_init function, those threads end up using the global
vairables in the vol file parser in an non-synchronized manner,
resulting in a seg-fault.

There is now a big lock around searches and additions from the mount
table in do_open. This lock granularity could be reduced.

Signed-off-by: Anand V. Avati <avati@dev.gluster.com>