extras: Add LD_PRELOAD test tool
This tool allows us to check the sanity of the LD_PRELOAD
mechanism so that we can be sure that an application's syscalls
will be redirected into booster when that library is LD_PRELOADed.

In case of failed syscalls, this tool should be run first
to see if the calls are redirected as required.

Signed-off-by: Anand V. Avati <avati@dev.gluster.com>

BUG: 251 (Improve booster debugging through ld-preload testing tool)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=251