distribute: Respect end-of-dir on readdir only for last subvol
We need to ensure that only the last subvolume's end-of-directory
notification is respected so that directory reading does not stop
before all subvolumes have been read. That could happen because the
posix for each subvolume sends a ENOENT on end-of-directory but in
distribute we're not concerned only with a posix's view of the
directory but the aggregated namespace' view of the directory,
which is maintained by distribute.

Signed-off-by: Shehjar Tikoo <shehjart@gluster.com>
Signed-off-by: Anand V. Avati <avati@dev.gluster.com>

BUG: 597 (miscellaneous fixes for xlators to work well with NFS xlator)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=597