core: Set ctx in global_xlator to prevent NULL ptr reference
This results in a crash in STACK_UNWIND after a call bail out.
Since the call bail does not occur in the context of an xlator,
the old_THIS in STACK_UNWIND is handed a global_xlator which did
not have its ctx set, resulting in a NULL de-reference.

Signed-off-by: Shehjar Tikoo <shehjart@gluster.com>
Signed-off-by: Anand V. Avati <avati@dev.gluster.com>

BUG: 844 (mixed testing with 35 clients)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=844