Misleading probe message for friend
Progress with probe if host already is present in the list and
state is not default or request received.

To test, probe from on host, on the other host, do a probe back to
host1. This should now continue with probe for the first time. Second
probe should return already friend msg.

Signed-off-by: shishir gowda <shishirng@gluster.com>
Signed-off-by: Vijay Bellur <vijay@dev.gluster.com>