syncdaemon: implement template substitutions for config values
So, for example, a log file setting of

  /var/log/${mastervol}/${eSlave}.log

will be substituted with the volume name of the master and the
canonicalized-escaped name of the slave for each master-slave pair.

As template expanders, beyond the various forms and derivatives
of master and slave, the following are also available:
 - gsyncd tunables (set in command line or in config)
 - for regexp sections, regexp group captures can be accessed
   via "match<i>_<n>", where i=1,2 corresponds to the i-th peer-rx
   in the section title and n=1,... to the n-th capture.

This will enable us to have a static configuration (not having to
add new entries on each gsyncd start).

Signed-off-by: Csaba Henk <csaba@lowlife.hu>
Signed-off-by: Anand Avati <avati@gluster.com>

BUG: 2785 (gsyncd logs on slave side go to /dev/null)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=2785