PUMP: perform opendir on pump xlator instead of its child
When replace-brick is performed, there is a high probability for the
directories on the source brick to not have any pending entry xattrs.
So for proper self-heal we need the force merge to kick in.
If the opendir is performed on pump xlator directory is examined and
if the entries on its children do not match then force merge is
triggered, so missing entries will be created on the sink.
Pending xattrs are set from source to sink on the files present on
source in this process. So when the lookup happens the self-heal
is triggered.

     Before this fix, the code is working because the self-heal source is
decided based on what file is biggest in size which is a wrong and
removed.

Signed-off-by: Pranith Kumar K <pranithk@gluster.com>
Signed-off-by: Anand Avati <avati@gluster.com>

BUG: 2586 (read child is set without checking the xattr)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=2586