glusterd: cleanup unneeded volumes after peer detach
Performs cleanup on the detached peer and in the cluster after a
peer detach, and also adds a new check before starting detach.

Cleanup -
 On the detached peer, cleanup removes the entries of those volumes
on the peer that do not have all their bricks on it. This prevents
these stale volumes from being added to a new cluster when peer is
attached to one.
 In the cluster, all those volumes which have all their bricks on the
detached peer are removed.

Checks-
 Checks if all the peers in the cluster are online and connected,
 except the peer being detached, before starting detach. Using force
 will bypass this check and do detach.

Change-Id: I4fef9ea3cc72ce8c4ce0a82b4ee8a1663a502061
BUG: 1926
Reviewed-on: http://review.gluster.com/431
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Krishnan Parthasarathi <kp@gluster.com>