changing of the volume types after volume is created.
with this patch, one can do all the below operations:

on a simple distribute volume 'test' with bricks A,B, add bricks C and D:

 bash# gluster volume add-brick test replica 2 C D

now the volume 'test' becomes distribute-replicate, with pairs A-C, B-D.
same holds good for stripe.

and now, one can change the replica count 2 on volume 'test' to 4 like below:

 bash# gluster volume add-brick test replica 4 E F G H

now, volume has 'replica' count 4, with pairs being 'A-C-E-F' and 'B-D-G-H'.

after this, if user wants to reduce the replica count, he can do

 bash# gluster volume remove-brick test replica 3 C H # (one from each pair)

This should work to reduce the replica count

Change-Id: I4f109141a9ba7be4a92a5f5810a4163baeb92973
BUG: 3305
Reviewed-on: http://review.gluster.com/158
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vijay@gluster.com>