Change the start/stop order of glusterd script at boot time.
In cases of client and server sharing the same machine, mounting GlusterFS at
boot time failed due to the start and stop order we use for glusterd. And people
had to resort to the idea of adding mount command in rc.local.

Currently the start and stop order at the boot time were 90 and 12, and netfs
started before `glusterd' and failed to mount GlusterFS. By changing
the rc order to 20 80 glusterd is started earlier than netfs, thus enabling
netfs to mount GlusterFS during boot process.

Thanks to: alex@net13.info for the suggestion in bug 2484.

Change-Id: I482f4dac5e5dd8bc5b8f5034c394f5dbe6c1fd3e
BUG: 2484
Reviewed-on: http://review.gluster.com/746
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Harshavardhana <fharshav@redhat.com>
Tested-by: Harshavardhana <fharshav@redhat.com>
Reviewed-by: Kaleb KEITHLEY <kkeithle@redhat.com>
Reviewed-by: Anand Avati <avati@gluster.com>