debug/error-gen: add option to generate errors at random percentage
Currently error-gen has a option where the percentage of the errors generated
can be specified. This patch adds a new option, which if given errors
are generated at random percentage. If both the random failure option and
the specific percentage of the errors are provided, then random error
generation is given preference.

Change-Id: I43fffcd7da8935761ed7d30e2a9677895d6a31fb
BUG: 767192
Signed-off-by: Raghavendra Bhat <raghavendrabhat@gluster.com>
Reviewed-on: http://review.gluster.com/786
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Jeff Darcy <jdarcy@redhat.com>
Reviewed-by: Vijay Bellur <vijay@gluster.com>