protocol/server: Do connection cleanup if reply fails
We observed that after the first connection cleanup happens on
DISCONNECT the lock calls in transit are granted or added in
blocked locks queue. These locks were never cleaned up after that
because no unlock would come up on that connection. This would
leave references on that transport so it would never be destroyed.
Now, the connection cleanup happens whenever the reply
submission fails.

Also cleaned up the old code which is not used any more.

Change-Id: Ie4fe6f388ed18d9c907cf8ae06b0b7fd0601a660
BUG: 765430
Signed-off-by: Pranith Kumar K <pranithk@gluster.com>
Reviewed-on: http://review.gluster.com/809
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Jeff Darcy <jdarcy@redhat.com>
Reviewed-by: Vijay Bellur <vijay@gluster.com>