cluster/afr: copy the parent's gfid from child loc while building parent loc
Suppose the process is not a fuse or nfs mounted client, and some other process
such as rebalance, then after lookups inode would not be linked to the inode
table (since the inode was created for rebalance purpose only), thus keeping
inode's gfid NULL.

And afr while building the parent loc using child loc, does not copy the
pargfid present in child'd loc structure. protocol/client will search for
the gfid either in loc or in loc->inode and assert if it cannot find the
gfid in either of them.

Change-Id: I882e449fb8b79d5c69e4a942abcd844dc4d5d30c
BUG: 799262
Signed-off-by: Raghavendra Bhat <raghavendrabhat@gluster.com>
Reviewed-on: http://review.gluster.com/2857
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Amar Tumballi <amarts@redhat.com>
Reviewed-by: Pranith Kumar Karampuri <pranithk@gluster.com>