protocol/client: Set grace_timer_needed flag to false on a disconnect.
A grace timer is registered on a disconnect, but a reconnect timer sends a
connect request after every 3sec and if the server is down, the client protocol
receives disconnect and a grace timer will be registered which on timeout will
increase the lk-version value. Its enough to register the grace timer once after
the first disconnect and later just ignore other psuedo disconnects until the
next successful reconnect.

Change-Id: I7f34a98a79fdc25370ef15a6712bfa316a9b7c7c
BUG: 807271
Signed-off-by: Mohammed Junaid <junaid@redhat.com>
Reviewed-on: http://review.gluster.com/3020
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vijay@gluster.com>