Optimize for small dicts, and avoid an overrun.
As dicts get used more and more in the I/O path (especially for xattrs and
the new xdata feature), removing some of their inherent inefficiency
becomes more important.  This patch addresses some of the issues around
allocating data_pair_t structures separately.  Along the way, I found that
the way we're allocating the "members" hash table was subtly wrong, and
could lead to a memory overrun.  This is a latent bug because nobody uses
dict_get_new_full that way, but I added an assert to guard against that
possibility.  One beneficial side effect is that we now save four pointers'
worth of space per dict, offsetting the extra space used for the new
members.

Change-Id: Ie8c3e49f1e584daec4b0d2d8ce9dafbc76fb57b2
BUG: 827448
Signed-off-by: Jeff Darcy <jdarcy@redhat.com>
Reviewed-on: http://review.gluster.com/3040
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Anand Avati <avati@redhat.com>