glusterfsd: don't ignore tmpfile write failure (2x)
In both glusterfs_volfile_reconfigure and mgmt_getspec_cbk,
we fwrite to a temporary file, fflush the stream, and then
rewind and read from it.  However, if either the fwrite or
fflush were to fail, we would have ignored a write failure
and would then read and process corrupt or incomplete input.
In each case, add a test of ferror so that we don't ignore
the write failure.

Found by inspection.

Change-Id: I8e84deb7d020a907870c9da2dde5c7371ce6ddb7
BUG: 789278
Signed-off-by: Jim Meyering <meyering@redhat.com>
Reviewed-on: http://review.gluster.com/3648
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Anand Avati <avati@redhat.com>