protocol/server: put the connection back if setvolume fails
In server_setvolume, put the connection back (thus removing it
from a the list of connections that protocol/server is maintaining
if the connection object is a newly created one) if some operations
such as getting the bound_xl etc fails. Otherwise the process might
segfault upon statedump, when it tries to access NULL bound_xl of a
connection.

Change-Id: Ifa340f534cd88d5601a9219da622b323fbe98b80
BUG: 858191
Signed-off-by: Raghavendra Bhat <raghavendra@redhat.com>
Reviewed-on: http://review.gluster.org/3953
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Amar Tumballi <amarts@redhat.com>
Reviewed-by: Anand Avati <avati@redhat.com>