cli,glusterd: Improve detach check validation
This patch improves the validation for the 'peer detach' command.

A check for if volumes exist with some bricks on the peer being detached
validation is added in peer detach code flow (even force would have this
validation).

This patch also gurantees that peer detach doesn't fail for a volume with all
its brick on the peer which is getting detached and there are no other bricks on
this peer.

The following steps need to be followed for removing a downed and unrecoverable
peer.
* If a replacement system is available
  - add it to the cluster
  - use replace-brick to migrate bricks of the downed peer to the new
    peer (since data cannot be recovered anyway use the 'replace-brick
    commit force' command)
  or,
  If no replacement system is available,
  - remove bricks of the downed peer using 'remove-brick'

Change-Id: Ie85ac5b66e87bec365fdedd8352b645bb25e1c33
BUG: 983590
Signed-off-by: Kaushal M <kaushal@redhat.com>
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-on: http://review.gluster.org/5325
Reviewed-by: Krishnan Parthasarathi <kparthas@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>