glusterd: add check in remove-brick start variant
The 'start' variant of the remove-brick command only applies at the dht
level wherein we can remove  all the bricks of a sub-volume (and remove
multiple such sub-volumes) but not select bricks of it.

This patch disallows removing individual replica bricks of multiple
sub-volumes (i.e. reducing the replcia count of the volume) using
remove-brick 'start'. The preferred method for such an operation is to use
commit force.

This patch also reverts the check to prevent removal of bricks from a
replicate volume (commit 0d415f7)

BUG: 961669
Change-Id: I447ad27f73a0963b5e09fb317bf7267a7a5a6147
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/5566
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Amar Tumballi <amarts@redhat.com>
Reviewed-by: Anand Avati <avati@redhat.com>