gsyncd / geo-rep: fix regular file creation on rename
On creation of a regular file on the slave (this happens
when an unlink()/rmdir() operation is performed on master
after a rename() or a link().

This patch fixes the regular file creation on the slave when
the above mentioned series of operation occurs. entry_ops()
on the slave for a rename() was missing the stat() information.

Change-Id: Idb80c6b3e27c1797dade1113ac8ce920db9c740c
BUG: 994353
Original Author: Venky Shankar <vshankar@redhat.com>
Signed-off-by: Venky Shankar <vshankar@redhat.com>
Reviewed-on: http://review.gluster.org/5509
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Avra Sengupta <asengupt@redhat.com>
Tested-by: Avra Sengupta <asengupt@redhat.com>
Reviewed-by: Anand Avati <avati@redhat.com>