Add the MAINTAINERS file
Different projects (Linux kernel, QEMU, ...) have a MAINTAINERS file in
the root of the git repository. This file mostly lists the maintainers
or main committers for certain files and directories. On occasion it
would be very helpful to know who to add as a reviewer for proposed
patches.

Change-Id: Ide68779c7477529808ccb0e7d421db160e3222ea
BUG: 1040351
URL: http://lists.nongnu.org/archive/html/gluster-devel/2013-11/msg00045.html
Signed-off-by: Niels de Vos <ndevos@redhat.com>
Reviewed-on: http://review.gluster.org/6480
Reviewed-by: Amar Tumballi <amarts@gmail.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Lalatendu Mohanty <lmohanty@redhat.com>
Reviewed-by: Kaleb KEITHLEY <kkeithle@redhat.com>