gfapi: Set log level to GF_LOG_NONE after closing log file in glfs_fini
Currently any logs that are logged after calling glfs_fini will
appear on stderr. Ideally to the end of the glfs_fini all the
operations will be complete and there should be no further logging.
But it appears that the client translator is still up and prints
some log messages even after calling glfs_fini.

In this patch the log level is set to GF_LOG_NONE once the log file
is closed. This fix is not required if wait for "CHILD_DOWN" event
is implemented in glfs_fini.

Change-Id: Idde80e9a89f3a5264d02b7e6fc2a84d67550b20a
BUG: 1054696
Signed-off-by: Poornima <pgurusid@redhat.com>
Reviewed-on: http://review.gluster.org/6726
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Anand Avati <avati@redhat.com>