glusterd/snapshot : Quorum check should not be made if we perform snapshot status command.
Problem : Snapshot status command used to fail as it used to
hit the quorum check path.

Solution : The condition checking where snapname is fetched
based on the presence of snap_volume is moved inside create
switch case.

And also moved the chunk of code which does the actual
quorum check to new function to make the code more
readable.

Change-Id: Idda2d7c576cdfab3a7d087bfa74bfa616372c20e
BUG: 1096700
Signed-off-by: Sachin Pandit <spandit@redhat.com>
Reviewed-on: http://review.gluster.org/7737
Reviewed-by: Rajesh Joseph <rjoseph@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Raghavendra Bhat <raghavendra@redhat.com>
Reviewed-by: Kaushal M <kaushal@redhat.com>