cli: 'Snapshot Volume: yes/no' for volume info needs to be removed
With initial design where the snap volume used to be displayed in
gluster volume info,
we used "Snap Volume: yes/on" to distinguish the volume whether its a
snap volume or the original volume.
But with new design the snap volumes are not listed in the volume info,
hence this entry (snap volume:) doesn't make sense to show.

Change-Id: Ic5b9948bf4ef74e89a611742c74a8989cb406866
BUG: 1098910
Signed-off-by: Vijaikumar M <vmallika@redhat.com>
Reviewed-on: http://review.gluster.org/7794
Reviewed-by: Avra Sengupta <asengupt@redhat.com>
Reviewed-by: Sachin Pandit <spandit@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Rajesh Joseph <rjoseph@redhat.com>
Reviewed-by: Krishnan Parthasarathi <kparthas@redhat.com>
Tested-by: Krishnan Parthasarathi <kparthas@redhat.com>