tests : Fix for spurious failure of bug-1104642.t.
Problem :
This test case checks whether global option is updated
when glusterd goes down and comes back up.

There might be a scenario where the handshake is not
completed yet and we issue a volume info in between.
In that case when we try to grep a key from gluster volume info
output it might not be there.

Hence to get some time we can issue a peer status command
and verify that the peer count is correct.

Change-Id: I2933d55ce8c80517a555fa3c2e4cd768cde30abf
BUG: 1104642
Signed-off-by: Sachin Pandit <spandit@redhat.com>
Reviewed-on: http://review.gluster.org/8041
Reviewed-by: Raghavendra Talur <rtalur@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>