glusterd/geo-rep: Check if the state file is present or not.
Currently every geo-rep operation other than start and stop
force, checks if the state-file is present or not before
performing the operation. This check is done to confirm that
the session on which the op is being performed is created.

This patch makes the start command also perform the same check
and accordingly display the correct error.

Change-Id: Ifd8bdfa3f362e91f58ead3202e2ebdbd411080e5
BUG: 1114469
Signed-off-by: Avra Sengupta <asengupt@redhat.com>
Reviewed-on: http://review.gluster.org/8202
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Kotresh HR <khiremat@redhat.com>
Reviewed-by: Aravinda VK <avishwan@redhat.com>
Reviewed-by: Krishnan Parthasarathi <kparthas@redhat.com>
Tested-by: Krishnan Parthasarathi <kparthas@redhat.com>