USS : Decide the log entry based on the type of graph.
Problem : When a lookup is issued, and if the entry is not found
then snapview-client will log failure stating that
"Lookup on normal graph failed with error Stale file handle"
irrespective of type of graph it received call back from.

Solution : Introduced a check to find out the graph from which
the snapview-client received call-back.

Change-Id: Iadd5b525c394be3675d40231711058e1cf1396cd
BUG: 1146479
Signed-off-by: Sachin Pandit <spandit@redhat.com>
Reviewed-on: http://review.gluster.org/8851
Reviewed-by: Raghavendra Bhat <raghavendra@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>