cluster/dht: Fix subvol check, to correctly determine cached file rename
The check to treat rename as a critical failure ignored when the cached
file is being renamed to new name, as the new name falls on the same
subvol as the cached file. This is in addition to when the target of the
rename does not exist.

The current change is simpler, as the rename logic, renames the cached
file in case the target exists and falls on the same subvol as source
name, OR the target does not exist and the hash of target falls on the
same subvol as source cached. These conditions mean we are renaming the
source, other conditions mean we are renaming the source linkto file
which we do not want to treat as a critical failure (and we also instruct
marker that it is an internal FOP and to not account for the same).

Change-Id: I4414e61a0d2b28a429fa747e545ef953e48cfb5b
BUG: 1161156
Signed-off-by: Shyam <srangana@redhat.com>
Reviewed-on: http://review.gluster.org/9063
Reviewed-by: N Balachandran <nbalacha@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: susant palai <spalai@redhat.com>
Reviewed-by: venkatesh somyajulu <vsomyaju@redhat.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>