heal: free leaked frames.
'gluster v heal <VOLNAME> info` waits for about 10 seconds before printing the
heal info and exiting.

Problem: glfsh_process_entries() is not freeing the frames it created.
Thus when glfs_fini() is called, it busy waits for 10 seconds for background
frames to finish before returning error.

Fix: Destroy the frames in that are created in glfsh_process_entries().

Change-Id: I5484d1b7301a355b913ae9c15b8a62471036c755
BUG: 1169335
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/9224
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Tested-by: Pranith Kumar Karampuri <pkarampu@redhat.com>