feature/changelog: Logging entry as well for explicit sync virtual xattr.
This is an improvement over the patch 'http://review.gluster.org/9337'
to trigger explicit geo-rep sync on regular files even if entry is not
present on the slave. An attempt is made to find the pargfid and
if available captures CREATE along with DATA in changelog.
CREATE is captured with default file permissions. Setting this virtual
setxattr on directories captures MKDIR in changelog. The value of
setxattr can be as follows.

If value = "1"       : Both CREATE and DATA is captured in changelog if
                       pargfid is available, else on DATA is captured.
   value = "any other: ENOTSUP is returned.

Usage:
  setfattr -n glusterfs.geo-rep.trigger-sync -v "1" <file-path>

NOTE: This patch supports explicit record of entries only for
      directories and regular files.

Change-Id: Iedde8b2c8bc3b78db524050d8c866ff664811d01
BUG: 1176934
Signed-off-by: Kotresh HR <khiremat@redhat.com>
Reviewed-on: http://review.gluster.org/9370
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Venky Shankar <vshankar@redhat.com>
Tested-by: Venky Shankar <vshankar@redhat.com>