cluster/dht: Don't restore entry when only one subvolume is present
Problem:
When rmdir fails with op_errno other than ENOENT/EACCES then self-heal
is attempted with zeroed-out stbuf. Only ia_type is filled from inode,
when the self-heal progresses, it sees that the directory is still
present and performs setattr with all valid flags set to '1' so the
file will be owned by root:root and the time goes to epoch

Fix:
This fixes the problem only in dht with single subvolume. Just don't
perform self-heal when there is a single subvolume.

Change-Id: I6c85b845105bc6bbe7805a14a48a2c5d7bc0c5b6
BUG: 1181367
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Reviewed-on: http://review.gluster.org/9435
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Shyamsundar Ranganathan <srangana@redhat.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>
Tested-by: Raghavendra G <rgowdapp@redhat.com>