test: Check till probe timeout, for conf change sync across glusterd
Currently the test case changed here, checks for the peer count to
be 1 until probe timeout and then checks for the changed
configuration, if it has been synced.

The peer count is not a gurantee that the configuration is also
in sync, hence changing this test case to check for the conf
update till probe timeout, by which time it should be in sync
(or at least that is our tolerance), and the test case deemed
as passing.

Change-Id: I4b1560979cfde3bd3bd691852d7d3a63e253bcf2
BUG: 1181203
Signed-off-by: Shyam <srangana@redhat.com>
Reviewed-on: http://review.gluster.org/9498
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-by: Krishnan Parthasarathi <kparthas@redhat.com>
Tested-by: Krishnan Parthasarathi <kparthas@redhat.com>