gluster-cli : Extra checks for word_count removed.
Coverity IDs:
1124453
1124454

In the first case, In the beginning of the function check is made
to see of the word_count is less than 4. If yes it returns. So it
wouldn't reach this part if that check takes the true path. So
this code is logically dead.

In second and third case, this is in false branch for word_cound <6.
So word_count can't possibly be less than 5 at this point.

Change-Id: Id5e1c81045ce7bc0ee6b7612ea54ef2b7f54b699
BUG: 789278
Signed-off-by: Nandaja Varma <nvarma@redhat.com>
Reviewed-on: http://review.gluster.org/9582
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Humble Devassy Chirammal <humble.devassy@gmail.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>