libgfapi: In glfs_init() set the cwd to "/".
Initially, when the cwd is NULL, while resolving a
relative path, it was taken to be relative from the root
of the volume.
But this behaviour changed with the patch:
http://review.gluster.org/#/c/9671/
with this patch, if cwd is NULL, all the fops on the files
(given relative path) under root directory will fail.

Hence, setting the cwd to "/" during glfs_init() itself.

Change-Id: I65018f6cd42539d09852069b84099034b9925418
BUG: 1196019
Signed-off-by: Poornima G <pgurusid@redhat.com>
Reviewed-on: http://review.gluster.org/9744
Reviewed-by: soumya k <skoduri@redhat.com>
Reviewed-by: Rajesh Joseph <rjoseph@redhat.com>
Reviewed-by: Niels de Vos <ndevos@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>