Tests: give sane default for NetBSD NFS mount retry count
Default NetBSD NFS retry count is 10000, which means tests will loop
for a long time if the server is not available.

We fix this by setting a default retry count to 2 (1 seems to low and
breaks regression). If mount_nfs is called with a retry option, it will
overrride this default.

BUG: 1129939
Change-Id: I1ae16f8caa74d6e9af1aa7a55fd111178af0ad78
Signed-off-by: Emmanuel Dreyfus <manu@netbsd.org>
Reviewed-on: http://review.gluster.org/9763
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Niels de Vos <ndevos@redhat.com>