glusterd: create nfs volfile even when NFS is disabled on all volumes
This is required to determine if gluster-nfs daemon needs to be
restarted. With http://review.gluster.org/9835 gluster-nfs volfile
wouldn't be created if all volumes had nfs disabled before they were
started even once. With the existing code, we wouldn't be able to
determine if gluster-nfs needs to be restarted or reconfigured without
the gluster-nfs volfile. This fix is ensure that we generate the
gluster-nfs volfile even if it wouldn't be started, to honour the above
requirement.

Change-Id: I86c6707870d838b03dd4d14b91b984cb43c33006
BUG: 1199944
Signed-off-by: Krishnan Parthasarathi <kparthas@redhat.com>
Reviewed-on: http://review.gluster.org/9851
Reviewed-by: Niels de Vos <ndevos@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-by: Kaushal M <kaushal@redhat.com>
Tested-by: Kaushal M <kaushal@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>