remove spurious failure test from bug-1087198.t
Consider below scenario in the quota xlator

T1 - write with delta1 bytes on fd1
     check_limit sees that delta1 bytes is not exceeding soft limit
T2 - write with delta2 bytes on fd1
     check_limit sees that delta2 bytes is not exceeding soft limit
T3 - delta1 and delta2 bytes are written to the disk.

Here delta1 and delta2 are checked separately and do not exceed
limit, but they together exceed the limit which is not checked.

We need to find a solution to solve this problem. Till then for
other regressions to pass, we remove the the test which checks for
soft limit crossed.

Change-Id: I8f76754e975c3315557a4c570db8bb5d9e56de15
BUG: 1202292
Signed-off-by: vmallika <vmallika@redhat.com>
Reviewed-on: http://review.gluster.org/9894
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>