tests: fix volume_exists to be used from EXPECT_WITHIN
Fixes the spurious volume-snapshot-clone.t regression failures.  In
brief, the problem is that the script wasn't waiting for config commands
to complete, and would *sometimes* query the status of a volume while
that volume was still being deleted.

It turns out that "!" doesn't work properly from EXPECT_WITHIN, so there
was a choice between changing that or changing volume_exists.  This
seemed less risky.  Because of code duplication, two instances of the
function had to be changed, and the other caller (volume-snapshot.t) did
too.

Change-Id: I766d4dc7c5b11038ede8e45d9d1f29cd02a622a0
BUG: 1163543
Signed-off-by: Jeff Darcy <jdarcy@redhat.com>
Reviewed-on: http://review.gluster.org/10053
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Kaleb KEITHLEY <kkeithle@redhat.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>