tier: relax libgfdb required version number
When calling dlopen() for libgfdb, do not specify the library version
number "libgfdb.so.0.0.1", since libtool will not always create libraries
or link with that name with the full 3-digit version. For instance on
NetBSD only up to the 2-digit version is available and "libgfdb.so.0.0.1"
does not exist.

Instead, just specify "libgfdb.so" and rely on smymlinks installed by
libtool to find the relevant library.

BUG: 1129939
Change-Id: I074b1009d3622a122fdaeb4b99658bca3277e211
Signed-off-by: Emmanuel Dreyfus <manu@netbsd.org>
Reviewed-on: http://review.gluster.org/10407
Tested-by: NetBSD Build System
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Vijay Bellur <vbellur@redhat.com>