glusterd: Invoking daemon reconfigure functions correctly
Problem : Scrub and bitd reconfigure functions were not invoking if
quota is not enabled.

Reason : In glusterd_svcs_reconfigure, if quota is not enabled then
it is returning in the middle of the function without calling bitd
and scrub reconfigure functions.

Fix : If quota is not enable on volume, skip quota reconfigure and
continue for other daemon reconfigure.

This patch also address the state dump issue for bitd and quotad
(logs the scrub and bitd info into state dump).

Change-Id: I39ea004b70c95543c08496245be595b3ea044a29
BUG: 1219355
Signed-off-by: anand <anekkunt@redhat.com>
Reviewed-on: http://review.gluster.org/10622
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Tested-by: NetBSD Build System
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Kaushal M <kaushal@redhat.com>