glusterd/ afr: remove the readdir-failover option from rebalance process
Remove the "replicate*.readdir-failover=off" xlator option from
rebalance process' runner arguments because this is the default behaviour in
afr-v2. i.e. reaaddir failover happens only if it is a new readdir
(offset=0). 'Continuing'readdirs are always wound on the same child
where the first readdir had happened and if that child is down, the
readdir fails.

The option cannot be removed from glusterd_volopt_map because there is a
chance of ending up in peer rejected state in a heterogenous cluster
(typical upgrade scenario) due to volinfo checksum mismatch. TODO:come up with
some warning message to the user for such deprecated options if a volume
set is attempted.

Change-Id: I6a33dc9fae7cccea8c5fd3d9de40d88d13ad3689
BUG: 1225328
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/10935
Reviewed-by: Krutika Dhananjay <kdhananj@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Reviewed-by: Anuradha Talur <atalur@redhat.com>
Reviewed-by: Krishnan Parthasarathi <kparthas@redhat.com>