libglusterfs: write error handling when filesystem have no space left
When no space left on filesystem and user want to perform any operation
which result to change in /var/lib/glusterd/* files then glusterd is
failing to write on temporary file.

Fix is to handle error when write failed on temporary files due to no
space left on file system.

Change-Id: I79d595776b42580da35e1282b4a3cd7fe3d14afe
BUG: 1226829
Signed-off-by: Gaurav Kumar Garg <ggarg@redhat.com>
Reviewed-on: http://review.gluster.org/11029
Reviewed-by: Niels de Vos <ndevos@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Cedric Buissart