gverify: Adding StrictHostKeyChecking=no for ssh verification
Before actually checking the compatibility between master and
slave, gverify checks if there is a passwordless ssh connection
between master to slave.  So if the entry of the slave was not
present in 'known_hosts' file in master gverify used to complain
that passwordless ssh has not been setup. This used to happen
even if there is a passwordless ssh between master to slave.

This patch fixes the above problem by using StrictHostKeyChecking=no
while doing ssh to slave.

Change-Id: I953c278e411ad6bc1dd1966ea6d895b05f890492
BUG: 1228696
Signed-off-by: M S Vishwanath Bhat <vbhat@redhat.com>
Reviewed-on: http://review.gluster.org/11106
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Aravinda VK <avishwan@redhat.com>