cluster/tier : fix for logical bugs/timing errors in tier.t
The md5sum fingerprints were not correctly compared after moving
files between the hot and cold tiers.

This version of tier.t uses a new function, "check_counters", to
ensure that the number of promotions/demotions is as expected.
This is intended to avoid spurious timing-related errors that were
seen with the old script.

Change-Id: I4a0ae7315493bfd307a0f68f21fa3ea33c88b08f
BUG: 1231268
Signed-off-by: Pamela Ousley <pousley@redhat.com>
Reviewed-on: http://review.gluster.org/11285
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Dan Lambright <dlambrig@redhat.com>