cluster/tier: fixes for migration over ec as cold tier
An opendir is done in rebalance. The graph constructed when
EC is used in tiering may have no local volumes (if
all the hot volumes are on one node and all the others on
another node). Previously the opendir only sent fops down
the local subvolumes for migration. They must be sent down
both the hot and cold subvolumes for tiering.

When setxattr2() received a NULL subvolume; this dereferenced
an uninitialized variable.

When a lookup is done during creation of the destination
file, the xattr dict is "polluted" with virtual xattrs.
These cause subsequent xattrs in the new file to not be
written by posix. They are required by EC.

The inode gfid for "entry_loc" in gf_defrag_migrate_single_file()
was not initialized. This made underlying translators
think the gfid was 0, and failed migration.

Change-Id: I6ccda8ca8e43485b9b354341bbfcb302496f632c
BUG: 1236212
Signed-off-by: Dan Lambright <dlambrig@redhat.com>
Reviewed-on: http://review.gluster.org/11433
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>
Reviewed-by: Shyamsundar Ranganathan <srangana@redhat.com>