tests: remove unwanted tests from volume-snapshot.t
volume-snapshot.t failspuriously because of having additional test cases
which restarts glusterd and they are really not needed as far as the test
coverage is concerned. Currently glusterd doesn't have a mechanism to indicate
that volumes handshaking has been completed or not, due to this even if the peer
handshaking finishes and all the peers are back to the cluster there could be a
case where any command which accesses the volume structure might end up in
corruption as volume handshaking is still in progress. This is because of volume
list is still not been made URCU protected.

Change-Id: Id8669c22584384f988be5e0a5a0deca7708a277d
BUG: 1255599
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-on: http://review.gluster.org/11972
Reviewed-by: Avra Sengupta <asengupt@redhat.com>
Reviewed-by: Raghavendra Talur <rtalur@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>