tests: Introduce a Vagrant VM based test environment
This introduces a mechanism using which a developer
could easily test the Gluster code in a VM environment.

Also, it will help bring uniformity in the environments
used by various developers.

How to use:
1. git checkout -b custom-branch-name
2. Make changes
3. Execute ./run-tests-in-vagrant.sh

What happens in the background:
1. A new directory is created:
    tests/vagrant/vagrant-custom-branch-name

   It will serve as the Vagrant dir which has the
   Vagrantfile and related ansible playbooks.

   The VM is started using Vagrant and provisioned
   using ansible.

2. The source dir is recursively copied over to the
   VM under /home/vagrant/glusterfs.

3. Gluster is source installed in VM.

What happens in the foreground:
1. run-tests.sh is executed in VM using ssh and output is displayed
   in the same terminal with option to use ctrl-c to interrupt the test
   midway. The VM would still persist and you could ssh into it.

Also, you can checkout a different branch elsewhere and execute
run-tests-in-vagrant.sh there to get another VM which would
execute tests on that code.

If you wish to make some changes in the code, you could:
a. Change the code in host and run the script again to repeat
   the whole process.
OR
b. vagrant ssh into the VM and make the changes in the VM.

Co-authored-by: Kaushal M <kaushal@redhat.com>
Co-authored-by: Michael Adam <obnox@samba.org>

Change-Id: Ic87801172c8b614cdecbdf2a765e1b3370a5faf7
BUG: 1291537
Signed-off-by: Michael Adam <obnox@samba.org>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>
Reviewed-on: http://review.gluster.org/12753
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Niels de Vos <ndevos@redhat.com>