tier/unlink: symlink failed to unlink
during unlink of a file, we will get stat just after
deleting the file, to see if the file is under migration or not.

but this stat call will fail for symlink if the actual file
was deleted.

So it is better not to send stat request from client if it is
a symlink as we are not migrating symlink.

Change-Id: Idc033b24fa3522b5261e579889d2195b43419682
BUG: 1293963
Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>
Reviewed-on: http://review.gluster.org/13074
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Reviewed-by: Dan Lambright <dlambrig@redhat.com>
Tested-by: Dan Lambright <dlambrig@redhat.com>