tests/quota : fix failing test on auxiliary mount point
In test file tests/bugs/quota/bug-1049323.t, test "EXPECT "0" get_aux"
fails in Fedora.

In get_aux function we grep for "/var/run/gluster/<volname>" to check
if auxiliary mount point is created and we return 0 on success else we
return 1. In fedora, auxiliary mount point is created on
"/run/gluster/<volname>". So it fails on Fedora.

The patch fixes it by just grepping for "/run/gluster/<volname>".

Change-Id: Icb59395df4a98109eaa8199cbdbdedcd1cbef27a
BUG: 1297740
Signed-off-by: Manikandan Selvaganesh <mselvaga@redhat.com>
Reviewed-on: http://review.gluster.org/13228
Reviewed-by: Niels de Vos <ndevos@redhat.com>
Reviewed-by: Raghavendra Talur <rtalur@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>