tier/libgfdb: Ordering query results from libgfdb
When querying we will order the query result to get
the hotest or the coldest files in the queried list
so that these files are migrated first.

Now here we are giving priority to the write heat(time and counters),
as it requires complex queries to have a composite
ordering of write and read + it has it impact on performance.

Change-Id: I2e0415dcfad4218b42c68fc5c2ed8d1f075ce9ea
Signed-off-by: Joseph Fernandes <josferna@redhat.com>
Reviewed-on: http://review.gluster.org/13607
Smoke: Gluster Build System <jenkins@build.gluster.com>
Tested-by: Joseph Fernandes
NetBSD-regression: NetBSD Build System <jenkins@build.gluster.org>
CentOS-regression: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Dan Lambright <dlambrig@redhat.com>