metrics: add code accross to get more xlator metrics
This commit is a step towards more inbuilt metrics from
inside of glusterfs library instead of depending on
io-stats translator to get more details.

This should help to plot graphs of glusterfs usage while
doing any operations across the translator graph. Also
this should work as a base to provide better internal details
to glusterfs translator graph.

There can be multiple work on top of this.

1. 'collectd' integration.
2. separate command from glusterd to get only relevant info
   to track performance metrics, instead of depending on
   statedump.
3. Tools to help debug performance issues, as each translator's
   each fop performance is measured by default.

Changes in this patch includes:
1. fop/cbk and failed_cbk counter per xlator->fops.
2. Change timeval to timespec for frame latency (for nanosecond resolution)
3. Make time/latency measurement default
4. Provide options in statedump and meta xlator to provide
   information about the metrics.

Updates #137

Change-Id: I4be3ee6278591c354d2fbb3d906b8cf91e4e1c72
Signed-off-by: Amar Tumballi <amarts@redhat.com>