scripts: Update rfc.sh to prompt for a github issue on RFE commits
If a commit does not contain a bug check with the committer if
this is an RFE (or such) and prompt to amend the commit with a
github issue that relates to the change.

Also, give some specifics on what string to add to refer to an
issue.

This change is meant to encourage committers to add github issues
and hence will not enforce the same, IOW the check can be ignored
and code can still be submitted. Reviewers would hence be
responsible to check for the same.

Change-Id: Id382247a787a96d55be6be554332a1b7ccde6bcd
BUG: 1428036
Signed-off-by: Shyam <srangana@redhat.com>
Reviewed-on: https://review.gluster.org/17032
Smoke: Gluster Build System <jenkins@build.gluster.org>
NetBSD-regression: NetBSD Build System <jenkins@build.gluster.org>
CentOS-regression: Gluster Build System <jenkins@build.gluster.org>
Reviewed-by: Niels de Vos <ndevos@redhat.com>