rfc.sh: make alignment check compulsory
Earlier, it was not reported if a function call had arguments on next
line and the args were not aligned to opening bracket. With this change,
it is a check that is displayed in output.

I have not analyzed what all --strict brings in. If it is found to be
more restrictive than required, we may choose to promote this *one* CHK
to WARN and stop using --strict.

Sample:

CHECK: Alignment should match open parenthesis
+                ret = gf_asprintf (&footer,
+                                " [%s]", strerror(errnum));

total: 0 warnings, 1 checks, 8 lines checked

Change-Id: I48834087efcae872b69746ba962b0bb1c130b5cd
BUG: 1469179
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>
Reviewed-on: https://review.gluster.org/17737
CentOS-regression: Gluster Build System <jenkins@build.gluster.org>
Smoke: Gluster Build System <jenkins@build.gluster.org>
Reviewed-by: Prashanth Pai <ppai@redhat.com>
Reviewed-by: ankitraj
Reviewed-by: Jeff Darcy <jeff@pl.atyp.us>
Reviewed-by: Sonal Arora <sarora@redhat.com>
Reviewed-by: Niels de Vos <ndevos@redhat.com>