mount/fuse : Fix parsing of vol_id for snapshot volume
For supporting sub-dir mount, we changed the volid. Which means anything
after a '/' in volume_id will be considered as sub-dir path.

But snapshot volume has vol_id stracture of /snaps/<volname>/<snapname>
which has to be considered as during the parsing.

Note 1: sub-dir mount is not supported on snapshot volume
Note 2: With sub-dir mount changes brick based mount for quota cannot be
        executed via mount command. It has to be a direct call via glusterfs

Change-Id: I0d824de0236b803db8a918f683dabb0cb523cb04
BUG: 1501235
Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>