libglusterfs/options: Add new member 'setkey' to xlator options
The 'setkey' will be used as the key by GD2 when setting the option
during volgen. 'setkey' also supports using varstrings.

This is mainly to be used for options, which use a different key for
'volume set' and in volfiles. For eg. the 'auth.*' options of
protocol/server.

The protocol/server xlator has been updated to make use of this for the
auth.allow and auth.reject options.

Updates #302

Change-Id: I1fd2fd69625c9db48595bd3f494c221625255169
Signed-off-by: Amar Tumballi <amarts@redhat.com>