rio: Added (s)setattr and (f)truncate support
(f)setattr was not handling dirty inodes, hence size
information returned was incorrect. This is fixed.

(f)truncate needs some special handling on the MDS when the
Op is just truncate, this is forwarded to the DS and then
returned back to the caller.

The changes are tested using glfsxmp program and also using
coreutils commands like truncate and stat.

Change-Id: I828ed7fdc88abb01237215352ec8714bc0c69b76
Signed-off-by: ShyamsundarR <srangana@redhat.com>