libglusterfs: export minimum necessary symbols
minimize risk of symbol collisions in global namespace.

see https://review.gluster.org/#/c/5697/ which Amar has
resurrected.

This is a strawman proposal to use an export-list to
only export the necessary symbols from libglusterfs. I suppose
some of this could be fixed by smarter use of static in the
function definitions.

It's a bit scary to see some of the names we expose. And then
there are the names we use in the reserved namespace.

One step short of going all the way to symbol versions

fixes gluster/glusterfs#382

Change-Id: Ifb848dfc655ef735dd27c73b7729e1188eb817f1
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>