client: make fuse direct I/O strategies explicit
So far the --direct-io-mode option has been presented
as of being Boolean valued. That is however not exact,
as a third behavior is chosen if the option is not
specified.

We accept now the "auto" value as an explicit choice
for the default heuristics, and indicate in the
descriptions of the option (which occur in commandline
help and in the gluterfs / mount.glusterfs man pages)
that auto is the default.

The default heuristics was briefly described in the
commandline help. We are getting rid of that, because:
- it's not the right place to provide such details;
- there is no guarantee of keeping the current heuristics
  so it might go out of sync with reality;
- that is already the case to some degree, because the
  description did not take into account that the default
  heuristics varies between platforms (on Mac, it's just
  "off"), and that xlators can also prescribe direct I/O
  for the file of their choice (see change
  I3fe3312cd96baa4eecfe1247ab7255b4f455f049).

Change-Id: Ia83479c0c67fe66b7fc2e0e8db5b7792d9f44b28
Signed-off-by: Csaba Henk <csaba@redhat.com>