fuse: make sure the send lookup on root instead of getattr()
This change was done in https://review.gluster.org/16945. While the
changes added there were required, it was not necessary to remove the
getattr part. As fuse's lookup on root(/) comes as getattr only, this
change is very much required.

The previous fix for this bug was to add the check for revalidation in
lookup when it was sent on root. But I had removed the part where
getattr is coming on root. The removing was not requried to fix the
issue then. Added back this part of the code, to make sure we have
proper validation of root inode in many places like acl, etc.

updates: bz#1437780
Change-Id: I859c4ee1a3f407465cbf19f8934530848424ff50
Signed-off-by: Amar Tumballi <amarts@redhat.com>