libglusterfs: Capture the dict response in syncop_xattrop_cbk
Problem:
Currently it is not possible to capture the xattrs values which
are set on the bricks by calling syncop_(f)xattrop, because the
response dict is not being assigned to any of the dictionaries.

Fix:
In the xattrop callback capture the response dict and send it
back to the caller if it is requested.

Change-Id: I9de9bcd97d6008091c9b060bcca3676cb9ae8ef9
fixes: bz#1572076
Signed-off-by: karthik-us <ksubrahm@redhat.com>