glusterd: Deadcode Coverity issue
This patch will fix coverity issue 74 from [1].

we are updating ret value line number 5011, and
immediately checking whether ret is having non
zero value at line number 5013.If ret is 0, then
only we continue to execute and we can reach line
number 5036. By the time we reach 5036, ret value
is always 0. So this block of code is redundant
here and removing it.

[1] https://download.gluster.org/pub/gluster/glusterfs/static-analysis/master/glusterfs-coverity/2018-07-23-5fa004f3/html/

Updates: bz#789278

Change-Id: Ia6e8ba2936e350f0d29a9151ab786622f5e750db
Signed-off-by: Oshank Kumar <okumar@redhat.com>