tests: Add ability to preserve older tarball for retried tests
When a test is retried, the cleanup directives overwrite the
older tarball with the latest one, thus losing the logs from
the failed run.

This patch changes run-tests.sh to rename the older tarball
when retrying a test, thus preserving the same.

The tarball is renamed using a time stamp and optionally a
trailing sequence number, in case the test fails within the
very second. Although the sequence # is not strictly required
as we retry only once, it provides a defence for any future
enhancements to the same.

Fixes: bz#1614062
Change-Id: I9afe486b0b6f6a26f2ad0642e38bc0ba15b3ecc9
Signed-off-by: ShyamsundarR <srangana@redhat.com>