rfc.sh: use 'git log -n1' instead of 'git show'
This is required mainly because if we use any format in a
documentation, and it has strings which are looking for commit
msg format (like fixes: bz...) in document, that too will be
considered in checks, even though it is not part of commit msg.

By using git log -n1, we would only check the content of commit
msg.

Updates: bz#1193929
Change-Id: I498590de6d3daa6be7bcbee1f083ef3fa6ecc96f
Signed-off-by: Amar Tumballi <amarts@redhat.com>