coverity: ignore tainted access reported in gf_free
Coverity reports tainted pointer access in _gf_free if the pointer passed in
was used by any IO related function by the caller. The taint within gf_free
is a false positive, as the tainted region is from the passed in pointer
till its allocated lenght, and not for contents before the pointer (i.e
the GF_MEM_HEADER_SIZE bytes before the passed in pointer), as that is
exclusively handled by the gf_alloc family of functions.

CID: 1228602, 1292646, 1292647, 1292648, 1292649, 1383192, 1383195, 1389691

Should additionally fix,
CID: 1292650, 1292651, 1357874, 1382373, 1382404, 1382407

Change-Id: I48c5a4028e7b0224c432bbc30f8c29408c2a466b
Updates: bz#789278
Signed-off-by: ShyamsundarR <srangana@redhat.com>