cli: add a warning/confirmation message in peer detach code path
On a multi node cluster if one of the node is detached which had active
clients mounted through the same server address, this can cause all the
clients to loose any volfile changes. This is due to the lack of infra
in glusterd to let client know the list of IPs and attempt to connect to
other active nodes as failback. Such framework does exist in GD2 but not
in GD1.

This patch ensures to take a preventive measure to have a warning
message thrown to user to ensure all such connected clients are
remounted through a different IP.

Change-Id: I740b01868abbd75bf0a609cfaf00114d4d78aa96
Fixes: bz#1647074
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>