xlators/mgmt/glusterd/src/glusterd-volgen.c: use dict_ new functions
In a previous patch (https://review.gluster.org/20769) we've
added the key length to be passed to dict_* funcs, to remove the need
to strlen() it. This patches makes use of these functions over
this whole file.

Please review carefully, as there are many many changes there.

Compile-tested only!
updates: bz#1193929
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>

Change-Id: I2e1ee340300ec330936c31becda6bfe1b6533281