Posix: Instead of using a thread for fsyncer, use synctask
With brick mux, the number of threads increases as the number of
bricks increases. As an initiative to reduce the number of
threads in brick mux scenario, replcing fsyncer thread to use
synctask infra.

afr_changelog_fsync is the only code that currently uses batch-fsync.
At the posix_fsync, the fsync fop is handed over to the fsyncer thread
if batch-fsync-mode is set(default set) and xdata contains batch-fsync.
Even when the batch_fsync_delay_usec is 0, the fsync is handed to the
fsyncer thread, so the main thread is not blocked on fsync. With this
patch this behaviour is changed. If the batch_fsync_delay_usec is 0,
there is no need to hand over fsync to separate thread, as fsync is
executed in io-threads. If batch_fsync_delay_usec > 0 then the batch
fsync happens on a timely basis in synctask thread.

Updates #475
Change-Id: I560ab1f3d690f353130cf46cbefec55fa8dfd3b4
Signed-off-by: Poornima G <pgurusid@redhat.com>