logging.c/h: aggressively remove sprintfs()
Try to reduce the number of sprintf() and string copies until we finally
log a log line.

Specifically, do not sprintf separately the timestr string and
do not sprintf/strcpy the appmsgstr separately - just stick it with
the header.
Hoping I did not leak anything or changed the log line formatting.

Also, allocate 4K (GF_LOG_BACKTRACE_SIZE) of memory
dynamically for trace output -
only if trace was actually requested (previously, it was
unconditionally)

In addition, some minor code formatting (unrelated to the above).

updates: bz#1193929
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>

Change-Id: Id2ccc85f9213a2b1c6eaa4a2f58ce043eac1824f