tests/bitrot: enable self-heal daemon before accessing the files
For testing the recovery of bad (or corrupted files) in a dispersed
volume, first enable self-heal daemon and let heal happen.

In bitrot feature, if a file becomes corrupted, the solution recommended
is to remove that file directly from the backend and then allowing heal
to happen. Hence turn on self-heal daemon and allow the heal to happen
after removing corrupted copy from the backend.

Change-Id: I7186110398ec1aee7e5727b9d1aac9a01db4d831
fixes: bz#1695327
Signed-off-by: Raghavendra Bhat <raghavendra@redhat.com>