simple-quota: support backend quota
This PR expects backend quota to properly return `statfs()` (or `df`) output.
If `features.simple-quota.use-backend` option is set, then there wont be any
active accounting in simple-quota translator, but only `setxattr()` and
`statfs()` is handled for special keys. We expect helper function to
set 'quota-limit' and also 'namespace' xattr to aid the glusterfs process in
general to identify the entries.

Change-Id: Id4229b720b57cde458b9b36e36ada3ffe2be0ac2
Signed-off-by: Amar Tumballi <amar@kadalu.io>